requirements.txt
django
pymysql
djangorestframework
django-cors-headers